Refactor a few core tests and add test for replacing url in demos
