+++ packages/@docfy/core/tests/__fixtures__/monorepo/packages/package1/components/button-demo/demo1.md
@@ -3,3 +3,5 @@ order: 1
 ---
 
 # Button demo
+
+[A link from demo](../../docs/intro.md)
